backend/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-CORS==5.0.1
Flask-JWT-Extended==4.7.1
psycopg[binary]
python-dotenv==1.1.0
marshmallow==3.25.1
gunicorn==23.0.0
beautifulsoup4>=4.12.0
pytest==8.3.5
pytest-flask==1.3.0
